Cleaning and moving Sequences parsing functions
